fix: Improve certificate verification and cipher suite selection
- Change length variables in extensions to `dp::usize` to prevent truncation issues
- Add `trusted_cas` and `expected_hostname` fields to `HandshakeConfig` for client certificate validation
- Prefer AES-GCM cipher suite if available, otherwise fall back to ChaCha20-Poly1305
- Configure record layer with selected cipher suite after handshake
- Implement full certificate parsing and signature verification using keylock
- Verify certificate validity times and hostname for client connections
- Extract raw Ed25519 public key from DER-encoded certificate
- Add `get_key_size` method to `RecordLayer` to support different key lengths
- Introduce `CipherSuite` enum and `set_cipher_suite` method in `RecordLayer`
- Implement AES-GCM encryption/decryption in `RecordLayer` when available and selected
- Add sequence number overflow checks and `MAX_SEQUENCE_NUMBER` constant to record layer
- Add `needs_key_update` method to `RecordLayer` and `TlsSession` to indicate when key updates are required
- Parse and log received TLS alerts, distinguishing `close_notify` from other alerts

+++ include/netpipe/tls/extensions.hpp
@@ -230,7 +230,7 @@ namespace netpipe::tls {
                 return dp::result::err(dp::Error::invalid_argument("key_share extension too short"));
             }
 
-            dp::u16 list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
+            dp::usize list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
             if (data.size() < 2 + list_length) {
                 return dp::result::err(dp::Error::invalid_argument("key_share extension data truncated"));
             }
@@ -297,7 +297,7 @@ namespace netpipe::tls {
                 return dp::result::err(dp::Error::invalid_argument("supported_groups extension too short"));
             }
 
-            dp::u16 list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
+            dp::usize list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
             if (data.size() < 2 + list_length || list_length % 2 != 0) {
                 return dp::result::err(dp::Error::invalid_argument("supported_groups extension malformed"));
             }
@@ -339,7 +339,7 @@ namespace netpipe::tls {
                 return dp::result::err(dp::Error::invalid_argument("signature_algorithms extension too short"));
             }
 
-            dp::u16 list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
+            dp::usize list_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
             if (data.size() < 2 + list_length || list_length % 2 != 0) {
                 return dp::result::err(dp::Error::invalid_argument("signature_algorithms extension malformed"));
             }
@@ -381,7 +381,7 @@ namespace netpipe::tls {
             return dp::result::err(dp::Error::invalid_argument("extensions list too short"));
         }
 
-        dp::u16 ext_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
+        dp::usize ext_length = (static_cast<dp::u16>(data[0]) << 8) | data[1];
         if (size < 2 + ext_length) {
             return dp::result::err(dp::Error::invalid_argument("extensions data truncated"));
         }
